Replace `isPushRegistered` with `isRegistered`
Introduce the method
`TextSecurePreferences.isRegistered(Context)`. Replace _almost_ all
calls to `TextSecurePreferences.isPushRegistered(Context)` with this
method; do the same for the method that sets this value.

Queries to whether or not the user is "push" registered are clearly
synonymous with determining their registered status. Since we are
introducing the ability for a user to be registered but not receive
push notifications this is ambiguous.

Minor changes have been made to surrounding code in order to provide
context as to which state is being requested.

Signed-off-by: Alex Melbourne <alex.melbourne@protonmail.com>