Demand a connection when not getting notifications
Include a check to determine if Play Services is present and thus Google Cloud
Messaging is being used when deciding if a connection is required. If we
aren't then `isConnectionNecessary()` needs to return `true`.

A messaging application isn't very good if it cannot receive
messages. By keeping our connection open we can continue to get data
in a timely fashion.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>