Check for services before creating phone account
Before proceeding to create a Red Phone account during registration
check if Google's Play Services are available; if they aren't skip
this step.

Because of the way Red Phone uses Google Cloud Messaging when trying
to connect to devices it is infeasible to support this feature if
the required technology is missing.

Losing encrypted calls is upsetting, but something is definitely
better than nothing.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>