Adapt to changes to `SignalServiceAccountManager`
Pass the extra required values to `verifyAccountWithCode`, and
`setAccountAttributes` to indicate that we desire to fetch messages
directly from the server if Google Play Services are not installed.

The interface to `SignalServiceAccountManager` has been modified so
that the methods `verifyAccountWithCode`, `verifyAccountWithToken`,
and `setAccountAttributes` contain a new trailing parameter:
`fetchesMessages`.

As an aside we also pass `voice` as `false` to these methods if
Play Services are unavailable because we have disabled Red Phone.

Signed-off-by: Daniel Micay <danielmicay@gmail.com>